Migrate TranslatableObject model into the core app
